requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.4
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.1
distro==1.4.0
Django==3.1.2
djangorestframework==3.12.1
filelock==3.0.12
flake8==3.8.4
html5lib==1.0.1
identify==1.5.6
idna==2.8
ipaddr==2.2.0
isort==5.6.4
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.3
pathspec==0.8.0
pep517==0.8.2
pre-commit==2.7.1
progress==1.5
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pyparsing==2.4.6
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.22.0
retrying==1.3.3
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.1.0
webencodings==0.5.1
